US9476 - Code Copy Icon Fix specs
